Move from dnsserver to plugin/pkg/watch
